+++ src/test/kotlin/racingcar/domain/numberpicker/RandomNumberPickerTest.kt
@@ -1,8 +1,7 @@
-package racingcar.domain
+package racingcar.domain.numberpicker
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.RepeatedTest
-import racingcar.domain.numberpicker.RandomNumberPicker
 
 class RandomNumberPickerTest {
     @RepeatedTest(10)
